Upstream version 1.1.0

# gpg: Signature made Tue 17 Nov 2015 16:42:56 GMT using RSA key ID F1C09EFB
# gpg: Good signature from "Alejandro Garrido Mota (mogaal) <garridomota@gmail.com>"
# gpg:                 aka "Alejandro Garrido Mota (mogaal) <mogaal@cpan.org>"
# gpg:                 aka "Alejandro Garrido Mota <alejandro@debian.org>"

+++ src/Progress.cpp
@@ -1,8 +1,7 @@
 ////////////////////////////////////////////////////////////////////////////////
 // Vramsteg - Utility for displaying progress bars in shell scripts.
 //
-// Copyright 2010 - 2013, Paul Beckingham, Federico Hernandez.
-// All rights reserved.
+// Copyright 2010 - 2015, Paul Beckingham, Federico Hernandez.
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -50,7 +49,7 @@ Progress::Progress ()
 }
 
 ////////////////////////////////////////////////////////////////////////////////
-Progress::Progress (const std::string& l, int w, int n, int x, bool p /* = true */, bool r /* = true */)
+Progress::Progress (const std::string& l, int w, long n, long x, bool p /* = true */, bool r /* = true */)
 : style ("")
 , label (l)
 , width (w)
@@ -89,13 +88,13 @@ void Progress::setWidth (int value)
 }
 
 ////////////////////////////////////////////////////////////////////////////////
-void Progress::setMin (int value)
+void Progress::setMin (long value)
 {
   minimum = value;
 }
 
 ////////////////////////////////////////////////////////////////////////////////
-void Progress::setMax (int value)
+void Progress::setMax (long value)
 {
   maximum = value;
 }
@@ -131,7 +130,7 @@ void Progress::showElapsed (bool value)
 }
 
 ////////////////////////////////////////////////////////////////////////////////
-void Progress::update (int value)
+void Progress::update (long value)
 {
   if (isatty (fileno (stdout)) && current != value)
   {
@@ -200,7 +199,7 @@ std::string Progress::formatTime (time_t t)
 void Progress::renderStyleDefault ()
 {
   // Fraction completed.
-  float fraction = (1.0 * (current - minimum)) / (maximum - minimum);
+  double fraction = (1.0 * (current - minimum)) / (maximum - minimum);
 
   // Elapsed time.
   time_t now = time (NULL);
@@ -282,7 +281,7 @@ void Progress::renderStyleDefault ()
 void Progress::renderStyleMono ()
 {
   // Fraction completed.
-  float fraction = (1.0 * (current - minimum)) / (maximum - minimum);
+  double fraction = (1.0 * (current - minimum)) / (maximum - minimum);
 
   // Elapsed time.
   time_t now = time (NULL);
@@ -364,7 +363,7 @@ void Progress::renderStyleMono ()
 void Progress::renderStyleText ()
 {
   // Fraction completed.
-  float fraction = (1.0 * (current - minimum)) / (maximum - minimum);
+  double fraction = (1.0 * (current - minimum)) / (maximum - minimum);
 
   // Elapsed time.
   time_t now = time (NULL);
